backend/requirements.txt
docling
langchain
torch
sentence-transformers
faiss-gpu
fastapi
uvicorn
pydantic
langchain_community
langchain_docling
langgraph
numpy<2
